Refactored code(divided to parts and corrected naming)
